refactor(comments): reshape into per-component barrel layout (pilot)
First feature to land the full per-component-barrel structure
defined in .claude/skills/project-structure.md:

  comments/
    CommentForm/{index.tsx, CommentForm.tsx}
    CommentItem/{index.tsx, CommentItem.tsx}
    CommentThread/{index.tsx, CommentThread.tsx}
    hooks/useComments.ts
    utils/
      commentActions.ts
      extractMentions.ts
      __tests__/extractMentions.test.ts

External consumers (e.g. ProgramSide importing CommentThread) need
no path changes — the barrel index.tsx re-exports preserve the
public path. Internal imports rewritten to ../utils/ and ../hooks/.

Validates the discipline at real scale before the rest of the
features migrate opportunistically.

Typecheck + 346 tests + build all pass.
