+++ src/features/comments/CommentForm/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CommentForm";
