+++ src/features/comments/CommentItem/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CommentItem";
