+++ src/features/comments/CommentThread/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CommentThread";
